+++ src/compute-engine/engine-declarations.ts
@@ -802,7 +802,8 @@ export function assignFn(
           ce,
           reconciled.type,
           declaredType,
-          effectsDeclared
+          effectsDeclared,
+          reconciled
         )
       )
         throw declaredTypeError(id, reconciled, declaredType);
@@ -911,7 +912,8 @@ export function assignFn(
             ce,
             reconciled.type,
             declaredType,
-            effectsDeclared
+            effectsDeclared,
+            reconciled
           )
         )
           throw declaredTypeError(id, reconciled, declaredType);
@@ -997,7 +999,8 @@ export function assignFn(
           ce,
           value.type,
           def.value.type,
-          def.value.effectsDeclared
+          def.value.effectsDeclared,
+          value
         )
       )
         throw declaredTypeError(id, value, def.value.type);